requirements.txt

---

# **File: requirements.txt**
```text
python-binance==1.0.16
python-dotenv==1.0.0
pytest==7.3.2
reportlab==4.0.3
